+++ components/FaucetForm.vue
@@ -1,0 +1,32 @@
+<template>
+<div class="faucetForm">
+    <b-col>
+        <b-row>
+        Faucet
+      </b-row>
+      <b-row>
+        Mosaic
+        <b-input></b-input>
+      </b-row>
+      <b-row>
+        Recipient
+        <b-input></b-input>
+      </b-row>
+      <b-row>
+        Amount
+        <b-input></b-input>
+      </b-row>
+      <b-row>
+        Submit
+        <b-button>CLAIM!</b-button>
+      </b-row>
+    </b-col>
+</div>
+</template>
+
+<style lang="scss" scoped>
+.faucetForm {
+    background-color: purple;
+    padding: 20px;
+}
+</style>
